+++ blogpessoal/src/main/java/org/generation/blogpessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package org.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -18,7 +19,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/tema")
@@ -41,7 +44,7 @@ public class TemaController {
 	}
 	
 	@GetMapping("/descricao/{descricao}")
-	public ResponseEntity<List<Tema>> getByNome(@PathVariable String descricao){
+	public ResponseEntity<List<Tema>> getByDescricao(@PathVariable String descricao){
 		return ResponseEntity.ok(temaRepository.findAllByDescricaoContainingIgnoreCase(descricao));
 	}
 	
@@ -52,7 +55,7 @@ public class TemaController {
 	}
 	
 	@PutMapping
-	public ResponseEntity<Tema> putTema(@RequestBody Tema tema){
+	public ResponseEntity<Tema> putTema(@Valid @RequestBody Tema tema){
 		
 		return temaRepository.findById(tema.getId())
 				.map(resposta -> ResponseEntity.status(HttpStatus.OK).body(temaRepository.save(tema)))
@@ -60,8 +63,14 @@ public class TemaController {
 	}
 	
 	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
-	public void deleteTema(@PathVariable Long id) {
+	public void deletePostagem(@PathVariable Long id) {
+		Optional<Tema> tema = temaRepository.findById(id);
+		
+		if(tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		
 		temaRepository.deleteById(id);
 	}
